chore: Depricate recentlyUsedWorkspaceId and recentlyUsedAppIds from user data
