+++ app/server/appsmith-server/src/main/java/com/appsmith/server/repositories/ce/CustomUserDataRepositoryCE.java
@@ -4,13 +4,11 @@ import com.appsmith.server.domains.UserData;
 import com.appsmith.server.repositories.AppsmithRepository;
 import reactor.core.publisher.Mono;
 
-import java.util.List;
-
 public interface CustomUserDataRepositoryCE extends AppsmithRepository<UserData> {
 
     Mono<Integer> saveReleaseNotesViewedVersion(String userId, String version);
 
-    Mono<Void> removeIdFromRecentlyUsedList(String userId, String workspaceId, List<String> applicationIds);
+    Mono<Void> removeIdFromRecentlyUsedList(String userId, String workspaceId);
 
     Mono<String> fetchMostRecentlyUsedWorkspaceId(String userId);
 }
