--- app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ApplicationFetcher.java
@@ -1,5 +1,0 @@
-package com.appsmith.server.solutions;
-
-import com.appsmith.server.solutions.ce.ApplicationFetcherCE;
-
-public interface ApplicationFetcher extends ApplicationFetcherCE {}
