+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/UserReleaseNotes.java
@@ -1,0 +1,5 @@
+package com.appsmith.server.solutions;
+
+import com.appsmith.server.solutions.ce.UserReleaseNotesCE;
+
+public interface UserReleaseNotes extends UserReleaseNotesCE {}
