+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ce/UserReleaseNotesCE.java
@@ -1,12 +1,8 @@
 package com.appsmith.server.solutions.ce;
 
 import com.appsmith.server.dtos.ReleaseItemsDTO;
-import com.appsmith.server.dtos.UserHomepageDTO;
 import reactor.core.publisher.Mono;
 
-public interface ApplicationFetcherCE {
-
-    Mono<UserHomepageDTO> getAllApplications();
-
+public interface UserReleaseNotesCE {
     Mono<ReleaseItemsDTO> getReleaseItems();
 }
